fix(transfer-detection): reactive post-dismiss banner; a11y id + semantic colors; Task 12 review
